+++ packages/forma-36-react-components/src/components/Autocomplete/Autocomplete.tsx
@@ -215,6 +215,7 @@ export const Autocomplete = <T extends {}>({
         <IconButton
           className={styles.inputIconButton}
           tabIndex={-1}
+          disabled={toggleProps.disabled}
           buttonType="muted"
           iconProps={{ as: toggleProps.query ? Close : ChevronDown }}
           onClick={toggleProps.onToggle}