+++ packages/forma-36-website/src/components/Layout.js
@@ -43,10 +43,6 @@ const Layout = (props) => {
               menuLinks {
                 name
                 link
-                menuLinks {
-                  name
-                  link
-                }
               }
             }
           }